requirements.txt
Flask==3.0.0
numpy
Pillow
gunicorn==21.2.0
requests==2.31.0
gtts==2.4.0
deep-translator==1.11.4
groq
# Updated for Render deployment compatibility
